requirements.txt
keyboard
requests
python-decouple